refactor: wallet context and hooks
