+++ apps/web/src/wallet/hooks/index.ts
@@ -1,13 +1,2 @@
-// hooks.ts
-import { useContext } from 'react';
-import { WalletContextType } from '../types';
-import { WalletContext } from '../provider';
-
-
-export function useWallet(): WalletContextType {
-  const context = useContext(WalletContext);
-  if (!context) {
-    throw new Error('useWallet must be used within a WalletProvider');
-  }
-  return context;
-}
+export * from './use-restore-wallet';
+export * from './use-wallet';
