+++ apps/web/src/wallet/index.ts
@@ -1,0 +1,3 @@
+export * from './hooks';
+export * from './types';
+export * from './provider';
